# app/static/swagger.yaml
swagger: "2.0"
info:
  title: "Mechanic API"
  description: "This API is used to create, update, and delete information for a mechanic shop"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: |
      Token Required!
      Please enter your JWT token **with `Bearer` prefix** like:
      ```
      Bearer "JWT TOKEN"
      ```
paths:
...
  /service_types:
    post:
      tags:
        - Service Types
      summary: "Create a service type that can be added for common services for cars."
      description: "Endpoint to create a service type for adding common service types for cars."
      consumes:
        - application/json
      produces:
        - application/json
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          description: "Service type to be created"
          required: true
          schema:
            $ref: "#/definitions/CreateServiceTypePayload"
      responses:
        201:
          description: "Successfully created a service type"
          schema:
            type: object
            properties:
              message:
                type: string
              service_type:
                $ref: "#/definitions/CreateServiceTypeResponse"
          examples:
            application/json:
              message: "Service Type added"
              service_type:
                id: 5
                name: Air Filter
                price: 18.99
        401:
          description: Unauthorized - Token missing or invalid
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: Forbidden - Admin access required
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    get:
      tags:
        - Service Types
      summary: "Gets all Service Types in database"
      description: "Endpoint to retrieve all Service Types with pagination support"
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: page
          in: query
          type: integer
          required: false
          description: "Page number (default: 1)"
        - name: per_page
          in: query
          type: integer
          required: false
          description: "Number of items per page (default: 10)"
      responses:
        200:
          description: "List of service types with pagination"
          schema:
            type: object
            properties:
              page:
                type: integer
              per_page:
                type: integer
              total_service_types:
                type: integer
              service_types:
                type: array
                items:
                  $ref: "#/definitions/CreateServiceTypeResponse"
          examples:
            application/json:
              page: 1
              per_page: 10
              total_service_types: 2
              service_types:
                - id: 1
                  description: null
                  name: Air Filter
                  price: 18.99
                - id: 2
                  description: null
                  name: Tire Replacement
                  price: 30.99
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /service_types/{id}:
    put:
      tags:
        - Service Types
      summary: "Update an existing service types"
      description: |
        Updates an existing type's data. Requires admin authentication.
      security:
        - bearerAuth: []
      consumes:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: "The ID of the service ticket to update"
        - in: body
          name: body
          description: Updated service type payload
          required: true
          schema:
            $ref: "#/definitions/CreateServiceTypePayload"
      responses:
        200:
          description: Successfully updated Service Type
          schema:
            type: object
            properties:
              message:
                type: string
              Service Type:
                $ref: "#/definitions/CreateServiceTicketPayload"
          examples:
            application/json:
              message: "Service Type updated"
              Service Type:
                id: 4
                description: null
                name: Fir Ailter
        400:
          description: Bad Request - Validation Error or Invalid service_type_ids
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: Unauthorized - Token missing or invalid
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: Forbidden - Admin access required
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Ticket not found
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      tags:
        - Service Types
      summary: "Endpoint to delete a Service Type"
      description: "Deletes a Service Type. Requires Admin token authentication, employee must be logged in to delete the ticket"
      consumes:
        - application/json
      produces:
        - application/json
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: "ID of the Service Type to delete"
      responses:
        200:
          description: "Service Type successfully deleted"
          schema:
            $ref: "#/definitions/DeleteServiceTypeResponse"
          examples:
            application/json:
              message: "Successfully deleted Service Type <service_type_id>"
        401:
          description: Unauthorized - Missing or invalid token
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: Forbidden - Admin token required
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Car not found
          schema:
            $ref: "#/definitions/ErrorResponse"
  /service_types/{service_type_id}/assign_service_type/{ticket_id}:
    put:
      tags:
        - Service Types
      summary: "Assign a Service Type to a ticket"
      description: "Assign an Service Type to a specific service ticket. Requires admin token."
      security:
        - bearerAuth: []
      parameters:
        - name: service_type_id
          in: path
          required: true
          type: integer
          description: "ID of the service type"
        - name: ticket_id
          in: path
          required: true
          type: integer
          description: "ID of the service type to assign"
      responses:
        200:
          description: "Service Type assigned to ticket successfully"
          schema:
            type: object
            properties:
              message:
                type: string
          examples:
            application/json:
              message: "Service Type 4 assign to ticket 7."
        401:
          description: Unauthorized - Missing or invalid token
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: Forbidden - Admin access required
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Ticket or Employee not found
          schema:
            $ref: "#/definitions/ErrorResponse"
  /service_types/{service_type_id}/remove_service_type/{ticket_id}:
    put:
      tags:
        - Service Types
      summary: "Remove a Service Type from a ticket"
      descriptions: "Removes an assigned service type from a service ticket. Requires Admin Token"
      security:
        - bearerAuth: []
      parameters:
        - name: service_type_id
          in: path
          required: true
          type: integer
          description: "ID of the service type"
        - name: ticket_id
          in: path
          required: true
          type: integer
          description: "ID of the ticket to remove from"
      responses:
        200:
          description: "Service Type removed from ticket successfully"
          schema:
            type: object
            properties:
              message:
                type: string
          examples:
            application/json:
              message: "Service Type 3 removed from ticket 7."
        401:
          description: Unauthorized - Missing or invalid token
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: Forbidden - Admin access required
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Ticket or Employee not found
          schema:
            $ref: "#/definitions/ErrorResponse"
definitions:
...
  ErrorResponse:
    type: "object"
    properties:
      message:
        type: "string"
...
  CreateServiceTicketPayload:
    type: object
    properties:
      id:
        type: integer
      VIN:
        type: string
      car_issue:
        type: string
      service_date:
        type: string
        format: date
      customer_id:
        type: integer
      car_id:
        type: integer
      is_major_damage:
        type: boolean
      services:
        type: array
        items:
          $ref: "#/definitions/ServiceType"
...
  CreateServiceTypePayload:
    type: object
    required:
      - name
      - price
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: number
        format: float
  CreateServiceTypeResponse:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: number
        format: float
      description:
        type: string
  ServiceType:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
  DeleteServiceTypeResponse:
    type: "object"
    properties:
      message:
        type: "string"